Added Create Order component. Updated NavBar. Fixed Order Controller.

+++ Models/Order.cs
@@ -8,7 +8,7 @@ namespace ShepherdsPie.Models
         public int Id { get; set; }
         public int? TableNumber { get; set; }
         public DateTime Date { get; set; }
-        public int TipAmount { get; set; }
+        public decimal TipAmount { get; set; }
 
         // Required (not optional) property for the user who took the order
         [Required]
